+++ Tank/GUI/Button.h
@@ -8,7 +8,7 @@
 class Button : public Label
 {
 public:
-	Button(std::string, sf::RenderWindow*, Coord*, std::string, int, int);
+	Button(std::string, sf::RenderWindow*, Coord*, std::string, float, float);
 	virtual ~Button();
 	virtual void draw();
 	virtual void onClick(void());
